src/main/java/com/xj/sell/service/impl/CategoryServiceImpl.java
package com.xj.sell.service.impl;

import com.xj.sell.dao.CategoryDao;
import com.xj.sell.pojo.ProductCategory;
import com.xj.sell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by xj
 * 2019-09-27 10:30
 **/
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    private CategoryDao dao;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return dao.findById(categoryId).orElse(null);
    }

    @Override
    public List<ProductCategory> findAll() {
        return dao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return dao.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory category) {
        return dao.save(category);
    }
}
